backend/requirements.txt
asgiref==3.6.0
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
joblib==1.2.0
numpy==1.21.6
scikit-learn==1.0.2
scipy==1.7.3
sqlparse==0.4.4
threadpoolctl==3.1.0
